requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
ipython==9.12.0
notebook==7.5.5
pandas==3.0.2